Structural refactor of validation, formatting and folders functions, some line-length cleanups.
